knowledge-os/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-os"
version = "0.1.0"
description = "AI Knowledge Graph Auto-Growth System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Knowledge OS Team"}
]
keywords = ["knowledge-graph", "rag", "langgraph", "mcp", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "chromadb>=0.5.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dateutil>=2.8.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "pre-commit>=3.0",
    "jupyter>=1.0",
]

[project.scripts]
knowledge-os = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
